requirements.txt
numpy
pillow
flask
flask-cors
numpy
werkzeug
matplotlib

--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.13.1

--extra-index-url https://download.pytorch.org/whl/cpu
torchvision